+++ dukan_data/src/commonMain/kotlin/net/thechance/mena/dukan/data/mapper/DukanMapper.kt
@@ -80,6 +80,6 @@ fun TopDiscountedDukanDto.toEntity(): TopDiscountedDukanPreview{
     return TopDiscountedDukanPreview(
         id = id,
         imageUrl = imageUrl,
-        discount = discount
+        discount = discount.toInt()
     )
 }
